#5 게시글 전체 조회 기능 추가(Read)
http://localhost:8088/user/post

엔드포인트 요청 결과 정상 작동

+++ src/main/java/vigo/com/viewgorithm/post/dto/PostDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Builder
 public class PostDto {
     private Long post_pk;
-    private int userPk;
+    private Long userPk;
     private String title;
     private String content;
     private LocalDateTime createdAt;
